+++ dartx/lib/dartx.dart
@@ -11,6 +11,7 @@ import 'package:collection/collection.dart' hide DelegatingList;
 import 'package:crypto/crypto.dart' as crypto;
 import 'package:quiver/collection.dart';
 
+part 'src/function.dart';
 part 'src/iterable_num.dart';
 part 'src/iterable.dart';
 part 'src/list.dart';