+++ dartx/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.2
 - Fixed docs
-- `'string'.reversed` now correctly handles grapheme clusters
+- Added `Function.curry()`, `Function.invoke()`, `Function.partial()` and `Function.flip()`
+- `'string'.reversed` now correctly handles grapheme clusters (like emoji)
 - **Breaking:** `'string'.chars` now returns the grapheme clusters instead of `Rune`s
 
 ## 0.1.1